services/auth/requirements.txt
# FastAPI and ASGI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
aiosqlite==0.19.0

# Redis
redis[hiredis]==5.0.8

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# MongoDB (for audit logs)
motor==3.6.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-cov==5.0.0

# Code Quality
ruff==0.6.9
black==24.8.0
mypy==1.11.2

# Utilities
python-dotenv==1.0.1
